@@ -1,13 +1,16 @@
 plugins {
     id 'java'
-    id 'org.springframework.boot' version '2.2.2.RELEASE'
-    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+    id 'org.springframework.boot' version '3.2.6'
+    id 'io.spring.dependency-management' version '1.1.4'
 }
 
 group 'com.example'
 version '1.0-SNAPSHOT'
 
-sourceCompatibility = 1.8
+java {
+    sourceCompatibility = JavaVersion.VERSION_21
+    targetCompatibility = JavaVersion.VERSION_21
+}
 
 repositories {
     mavenCentral()
@@ -15,6 +18,8 @@ repositories {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    compile group: 'net.logstash.logback', name: 'logstash-logback-encoder', version: '5.2'
-    testCompile group: 'junit', name: 'junit', version: '4.12'
+    implementation 'net.logstash.logback:logstash-logback-encoder:7.4' 
+    implementation 'jakarta.annotation:jakarta.annotation-api:2.1.1'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
+
